Fixed different signedness comparisons.

+++ src/operations/spatial_partitions.hpp
@@ -67,7 +67,7 @@ basis::field<basis::real_space, int> voronoi_field(std::vector<vector3<double, c
 template <typename CellType>
 basis::field_set<basis::real_space, int> local_radii_field(std::vector<vector3<double, cartesian>> const & local_centers, std::vector<double> const & local_radii, CellType const & cell, basis::real_space const & bas){
     auto nloc = static_cast<int>(local_centers.size());
-    assert(nloc == local_radii.size());
+    assert(nloc == (long long) local_radii.size());
 
     basis::field_set<basis::real_space, int> local_field(bas, nloc);
     local_field.fill(0);
